Modified extract_twop_mesons_h5.sh to support the High Momenta Form

+++ scripts/twop/extract_twop_mesons_h5.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-if [ $# -ne 10 ]
+if [ $# -ne 11 ]
 then
-    echo "Usage: $0 <extract_dir> <h5_file-prefix> <out_dir> <src_list> <mom_list> <conf> <T> <Qsq> <exe_dir> <meson, see list below"
+    echo "Usage: $0 <extract_dir> <h5_file-prefix> <out_dir> <src_list> <mom_list> <conf> <T> <Qsq> <exe_dir> <HighMomForm:1 LowMomForm:0> <meson, see list below"
     printf "  0: all\n  1: pseudoscalar\n  2: scalar\n  3: g5g1\n  4: g5g2\n  5: g5g3\n  6: g5g4\n  7: g1\n  8: g2\n  9: g3\n 10: g4\n"
     exit
 fi
@@ -16,7 +16,8 @@ CONF=$6
 T=$7
 Qsq=$8
 EXE_DIR=$9
-mes=${10}
+HighMomForm=${10}
+mes=${11}
 
 LIST[1]="pseudoscalar"
 LIST[2]="scalar"
@@ -40,29 +41,67 @@ then
 fi
 
 
-while read x y z t
-do 
-    FILE=${DIR}/${H5_PRE}.${CONF}_mesons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
+if [ ${HighMomForm} -eq 0 ]
+then
 
-    src=sx${x}sy${y}sz${z}st${t}
-    EX_DIR=${DIR}/${src}
-    mkdir -p ${EX_DIR}
+    while read x y z t
+    do 
+	FILE=${DIR}/${H5_PRE}.${CONF}_mesons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
 
-    OUT_FILE=${OUT_DIR}/twop.${CONF}.mesons.SS.${x}.${y}.${z}.${t}.dat
+	src=sx${x}sy${y}sz${z}st${t}
+	EX_DIR=${DIR}/${src}
+	mkdir -p ${EX_DIR}
 
-    rm -f ${OUT_FILE}
+	OUT_FILE=${OUT_DIR}/twop.${CONF}.mesons.SS.${x}.${y}.${z}.${t}.dat
 
-    for tp in ${MESON_LIST}
-    do
-	while read mx my mz
+	rm -f ${OUT_FILE}
+
+	for tp in ${MESON_LIST}
 	do
-	    EX_FILE=${EX_DIR}/twop.${CONF}.mesons.${tp}.${mx}_${my}_${mz}.SS.${x}.${y}.${z}.${t}.dat
+	    while read mx my mz
+	    do
+		EX_FILE=${EX_DIR}/twop.${CONF}.mesons.${tp}.${mx}_${my}_${mz}.SS.${x}.${y}.${z}.${t}.dat
+		
+		${EXE_DIR}/extract_twop_mesons_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${CONF} ${src} ${T} 0
+
+		cat ${EX_FILE} >> ${OUT_FILE}
+
+	    done < ${MOM_LIST}
+	done
+
+	rm -rf ${EX_DIR}
+
+    done < ${SRC_LIST}
+
+fi
+
+
+if [ ${HighMomForm} -eq 1 ]
+then
+
+    while read x y z t
+    do 
+	FILE=${DIR}/${H5_PRE}.${CONF}_mesons_Qsq${Qsq}_SS.${x}.${y}.${z}.${t}.h5
+
+	src=sx${x}sy${y}sz${z}st${t}
+	EX_DIR=${DIR}/${src}
+	mkdir -p ${EX_DIR}
+
+	OUT_FILE=${OUT_DIR}/twop.${CONF}.mesons.SS.${x}.${y}.${z}.${t}.dat
+
+	rm -f ${OUT_FILE}
+
+	for tp in ${MESON_LIST}
+	do
+	    EX_FILE=${EX_DIR}/twop.${CONF}.mesons.${tp}.SS.${x}.${y}.${z}.${t}.dat
 	    
-	    ${EXE_DIR}/extract_twop_mesons_h5 ${FILE} ${EX_FILE} ${tp} ${mx} ${my} ${mz} ${CONF} ${src} ${T} 0
+	    ${EXE_DIR}/extract_twop_mesons_HighMomForm_h5 ${FILE} ${EX_FILE} ${tp} ${CONF} ${src} ${T} 0
+	    
+	    cat ${EX_FILE} >> ${OUT_FILE}
+	done
 
-            cat ${EX_FILE} >> ${OUT_FILE}
+	rm -rf ${EX_DIR}
 
-	done < ${MOM_LIST}
-    done
+    done < ${SRC_LIST}
 
-done < ${SRC_LIST}
+fi
